fix: adapt codetracer's vscode.nim/build/extension so it can load it
TODO: load in webviews, restore DAP/others
TODO: eventually a better way to build?
